# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Protocol Banks API
  version: 1.0.0
  description: |
    Enterprise-grade crypto payment infrastructure API.
    Non-custodial, multi-chain payment solutions with Aave/JustLend yield integration.

    ## Authentication
    All authenticated endpoints require the `x-user-address` header containing a valid wallet address (EVM 0x... or TRON T...).

    ## Rate Limiting
    - Default: 60 requests/minute
    - Yield endpoints: 30 requests/minute
    - Payment mutations: 10 requests/minute

    Rate limit headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`
  contact:
    name: Protocol Banks Support
    url: https://protocolbanks.com
  license:
    name: MIT
servers:
  - url: https://app.protocolbanks.com/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
components:
  securitySchemes:
    WalletAuth:
      type: apiKey
      in: header
      name: x-user-address
      description: Wallet address (EVM 0x... or TRON T...)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for programmatic access
  schemas:
...
    YieldBalance:
      type: object
      properties:
        merchant:
          type: string
        network:
          type: string
        principal:
          type: string
        interest:
          type: string
        totalBalance:
          type: string
        apy:
          type: number
        protocol:
          type: string
          enum: [Aave V3, JustLend]
        networkType:
          type: string
          enum: [EVM, TRON]
    CrossNetworkSummary:
      type: object
      properties:
        totalPrincipal:
          type: string
        totalInterest:
          type: string
        totalBalance:
          type: string
        averageAPY:
          type: number
        balances:
          type: array
          items:
            $ref: '#/components/schemas/YieldBalance'
...
security:
  - WalletAuth: []
paths:
...
  /verify-payment:
    post:
      tags: [Payments]
      summary: Verify payment completion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [txHash, orderId, amount]
              properties:
                txHash:
                  type: string
                orderId:
                  type: string
                amount:
                  type: string
      responses:
        '200':
          description: Payment verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  reason:
                    type: string

  # ============= Vendors =============
...
  /yield/balance:
    get:
      tags: [Yield]
      summary: Get yield balance
      parameters:
        - name: merchant
          in: query
          required: true
          schema:
            type: string
          description: Merchant wallet address (EVM or TRON)
        - name: network
          in: query
          schema:
            type: string
            enum: [ethereum, base, arbitrum, tron, tron-nile]
        - name: summary
          in: query
          schema:
            type: string
            enum: ['true', 'false']
          description: Set to 'true' for cross-network summary
      responses:
        '200':
          description: Yield balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/YieldBalance'
                      - $ref: '#/components/schemas/CrossNetworkSummary'
        '400':
          description: Invalid parameters
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded
  /yield/stats:
    get:
      tags: [Yield]
      summary: Get yield statistics
      parameters:
        - name: network
          in: query
          schema:
            type: string
            enum: [ethereum, base, arbitrum, tron, tron-nile]
      responses:
        '200':
          description: Yield statistics
        '429':
          description: Rate limit exceeded
...